Revert "Backend refactored and default db changes to postgres. (#37)"
This reverts commit c589440cebeafa7b5aa0bca29c5f319849092a34.
