+++ .env.example
@@ -1,23 +1,14 @@
 DJANGO_ADMIN_URL=admin/
 DJANGO_SECRET_KEY=secret_key
 DEBUG=true
-
 AWS_ACCESS_KEY_ID=id
 AWS_SECRET_ACCESS_KEY=key
 AWS_STORAGE_BUCKET_NAME=bucket
 AWS_S3_REGION_NAME=region
-
 SENDER_EMAIL=email
 EMAIL_HOST=host
 EMAIL_PORT=port
 EMAIL_HOST_USER=user
 EMAIL_HOST_PASSWORD=password
-
-CELERY_BROKER_URL=amqp://guest:guest@localhost:5672//
-CELERY_RESULT_BACKEND=db+postgresql://postgres:postgres@localhost:5432/djangoindia-local
-
-DATABASE_NAME=djangoindia-local
-DATABASE_USER=postgres
-DATABASE_PASSWORD=postgres
-DATABASE_HOST=localhost
-DATABASE_PORT=5432
+CELERY_BROKER_URL=amqp://localhost
+CELERY_RESULT_BACKEND=rpc://
