--- backend/djangoindia/api/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class ApiConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'djangoindia.api'
