--- backend/djangoindia/api/serializers/__init__.py
@@ -1,1 +1,0 @@
-from .event import EventSerializer, EventRegistrationSerializer
