--- backend/djangoindia/api/urls/__init__.py
@@ -1,3 +1,0 @@
-from .event import urlpatterns as event_urls
-
-urlpatterns = event_urls
