--- backend/djangoindia/api/views/__init__.py
@@ -1,1 +1,0 @@
-from .event import EventAPIView
