--- backend/djangoindia/bg_tasks/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class BgTasksConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'djangoindia.bg_tasks'
