--- backend/djangoindia/db/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class DbConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "djangoindia.db"
