--- backend/djangoindia/db/models/__init__.py
@@ -1,1 +1,0 @@
-from djangoindia.db.models.event import Event, EventRegistration
