+++ backend/djangoindia/settings/base.py
@@ -48,11 +48,10 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "celery",
-    "corsheaders",
-    "djangoindia.api",
-    "djangoindia.bg_tasks",
-    "djangoindia.db",
     "rest_framework",
+    "storages",
+    "events",
+    "corsheaders",
 ]
 
 MIDDLEWARE = [
@@ -93,12 +92,8 @@ WSGI_APPLICATION = "djangoindia.wsgi.application"
 
 DATABASES = {
     "default": {
-        "ENGINE": "django.db.backends.postgresql",
-        "NAME": os.getenv("DATABASE_NAME"),
-        "USER": os.getenv("DATABASE_USER"),
-        "PASSWORD": os.getenv("DATABASE_PASSWORD"),
-        "HOST": os.getenv("DATABASE_HOST"),
-        "PORT": os.getenv("DATABASE_PORT"),
+        "ENGINE": "django.db.backends.sqlite3",
+        "NAME": BASE_DIR / "db.sqlite3",
     }
 }
