+++ backend/djangoindia/settings/local.py
@@ -1,3 +1,5 @@
+import os
+
 from .base import *  # noqa: F403, F401
 
 EMAIL_BACKEND = "django.core.mail.backends.console.EmailBackend"
