+++ backend/djangoindia/settings/production.py
@@ -1,7 +1,8 @@
+import os
+
 from .base import *  # noqa: F403, F401
 
 EMAIL_BACKEND = "django.core.mail.backends.smtp.EmailBackend"
-INSTALLED_APPS+=["storages"]
 
 # EMail settings
 EMAIL_HOST = os.environ.get("EMAIL_HOST")
