+++ backend/events/serializers.py
@@ -1,0 +1,55 @@
+from rest_framework import serializers
+
+from .models import Event, EventRegistration
+
+
+class EventSerializer(serializers.Serializer):
+    id = serializers.UUIDField(read_only=True)
+    name = serializers.CharField(max_length=100)
+    description = serializers.CharField()
+    cover_image = serializers.ImageField()
+    venue= serializers.CharField()
+    city= serializers.CharField()
+    venue_map_link= serializers.URLField()
+    date_time= serializers.DateTimeField()
+
+    def create(self, validated_data):
+        return Event.objects.create(**validated_data)
+
+    def update(self, instance, validated_data):
+        instance.name = validated_data.get("name", instance.name)
+        instance.description = validated_data.get("description", instance.description)
+        instance.save()
+        return instance
+
+
+class EventRegistrationSerializer(serializers.Serializer):
+    event = serializers.PrimaryKeyRelatedField(queryset=Event.objects.all())
+    email = serializers.EmailField()
+    first_name = serializers.CharField(max_length=255)
+    last_name = serializers.CharField(max_length=255)
+    occupation = serializers.ChoiceField(choices=EventRegistration.OCCUPATION_CHOICES)
+    gender = serializers.ChoiceField(choices=EventRegistration.GENDER_CHOICES)
+    linkedin = serializers.URLField()
+    github = serializers.URLField(required=False, allow_blank=True)
+    twitter = serializers.URLField(required=False, allow_blank=True)
+    other_links = serializers.URLField(required=False, allow_blank=True)
+    rsvp = serializers.BooleanField(default=False)
+
+    def create(self, validated_data):
+        return EventRegistration.objects.create(**validated_data)
+
+    def update(self, instance, validated_data):
+        instance.event = validated_data.get("event", instance.event)
+        instance.email = validated_data.get("email", instance.email)
+        instance.first_name = validated_data.get("first_name", instance.first_name)
+        instance.last_name = validated_data.get("last_name", instance.last_name)
+        instance.occupation = validated_data.get("occupation", instance.occupation)
+        instance.gender = validated_data.get("gender", instance.gender)
+        instance.linkedin = validated_data.get("linkedin", instance.linkedin)
+        instance.github = validated_data.get("github", instance.github)
+        instance.twitter = validated_data.get("twitter", instance.twitter)
+        instance.other_links = validated_data.get("other_links", instance.other_links)
+        instance.rsvp = validated_data.get("rsvp", instance.rsvp)
+        instance.save()
+        return instance
